Dropped the int in $int() somehow in my edit. Oops.

+++ cligen/humanUt.nim
@@ -137,7 +137,7 @@ proc humanDuration*(dt: int, fmt: string, plain=false): string =
       let dec = parseInt(div_dec[1])
       dts = formatFloat(dt.float / parseInt(div_dec[0]).float, ffDecimal, dec)
     else:
-      dts = $(dt.float / parseInt(cols[0]).float)
+      dts = $int(dt.float / parseInt(cols[0]).float)
     if cols.len > 2: result.add textAttrOn(cols[2..^1], plain)
     result.add dts
     if cols[1].startsWith('<'):
